% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm for greatest‑common‑divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{numeric scalar, first input integer}

\item{b}{numeric scalar, second input integer}
}
\value{
numeric scalar, greatest common divisor of a and b
}
\description{
Compute greatest common divisor of two integers using Euclidean algorithm.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100,1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
