% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia graph dataset}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Starting vertex}
  \item{v2}{Ending vertex}
  \item{w}{Weight of the edge}
}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
Wikipedia graph dataset
}
\keyword{datasets}
